Delete key-{import|export}-common
