+++ Cargo.toml
@@ -1,9 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
-  "./key-import-common",
   "./key-import-client",
-  "./key-export-common",
   "./key-export-client",
   "./core",
 ]
