+++ core/Cargo.toml
@@ -8,7 +8,11 @@ publish = false
 generic-ec = { version = "0.2", default-features = false, features = ["serde", "curve-secp256k1", "curve-stark", "curve-ed25519"] }
 
 rand_core = { version = "0.6", default-features = false }
+
 serde = "1"
+serde_with = { version = "1", features = ["base64"], optional = true }
+hex = { version = "0.4", default-features = false, features = ["alloc", "serde"], optional = true }
+
 base64 = "0.21"
 aes-gcm = "0.10"
 hkdf = "0.12"
@@ -27,3 +31,5 @@ serde_json = "1"
 
 [features]
 std = []
+import = ["dep:serde_with"]
+export = ["dep:serde_with", "dep:hex"]
