--- key-import-common/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "dfns-key-import-common"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dfns-trusted-dealer-core = { path = "../core" }
-generic-ec = { version = "0.2", default-features = false, features = ["serde", "curve-secp256k1"] }
-generic-ec-zkp = { version = "0.2", default-features = false, features = ["serde", "alloc"] }
-
-rand_core = { version = "0.6", default-features = false }
-serde = "1"
-serde_with = { version = "1", features = ["base64"] }
-hex = { version = "0.4", default-features = false, features = ["alloc", "serde"]}
-
-aes-gcm = "0.10"
-hkdf = "0.12"
-sha2 = "0.10"
-
-[dev-dependencies]
-key-share = { version = "0.1", features = ["spof"] }
-rand_dev = "0.1"
-serde_json = "1"
-
-[features]
-std = ["dfns-trusted-dealer-core/std"]
